refactor: extract preview and install behavior
