+++ apps/web/src/App.tsx
@@ -1,34 +1,24 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 
 import type {
   CreateRoomResponse,
   JoinRoomResponse,
   LobbyRequestStatusResponse,
   QualityPreset,
-  RequestedMedia,
 } from "@lowtime/shared";
 
 import { CallPage } from "./features/call/call-page.js";
 import { useCallFlow } from "./features/call/call-effects.js";
+import { useInstallPrompt } from "./features/home/install-effects.js";
 import { HomePage } from "./features/home/home-page.js";
-import { joinRoomRequest, startPreviewRequest, submitLobbyAction } from "./features/room/room-actions.js";
+import { joinRoomRequest, submitLobbyAction } from "./features/room/room-actions.js";
+import { useDevicePreview } from "./features/room/preview-effects.js";
 import { useRoomPageData } from "./features/room/room-effects.js";
 import { RoomPage } from "./features/room/room-page.js";
 import { useWaitingRoomState } from "./features/waiting/waiting-effects.js";
 import { WaitingPage } from "./features/waiting/waiting-page.js";
 import { assessNetworkHealth, type NetworkHealth } from "./network-health.js";
 import {
-  stopMediaStream,
-  type PreviewState,
-} from "./device-preview.js";
-import {
-  attachInstallPromptListeners,
-  isPwaInstalled,
-  promptForInstallation,
-  type BeforeInstallPromptEvent,
-} from "./pwa.js";
-import {
-  buildRequestedMedia,
   clearStoredLobbyRequest,
   getApiBaseUrl,
   getCallRoute,
@@ -42,11 +32,6 @@ import {
 } from "./room-entry.js";
 
 const DEFAULT_QUALITY_PRESET: QualityPreset = "balanced";
-const DEFAULT_REQUESTED_MEDIA: RequestedMedia = {
-  audio: true,
-  video: true,
-};
-
 export function App() {
   const [viewState, setViewState] = useState(() => getViewState(window.location.pathname));
   const [createResult, setCreateResult] = useState<CreateRoomResponse | null>(null);
@@ -55,10 +40,6 @@ export function App() {
 
   const [displayName, setDisplayName] = useState("");
   const [selectedQualityPreset, setSelectedQualityPreset] = useState<QualityPreset>(DEFAULT_QUALITY_PRESET);
-  const [previewAudioEnabled, setPreviewAudioEnabled] = useState(DEFAULT_REQUESTED_MEDIA.audio);
-  const [previewVideoEnabled, setPreviewVideoEnabled] = useState(DEFAULT_REQUESTED_MEDIA.video);
-  const [previewState, setPreviewState] = useState<PreviewState>("idle");
-  const [previewError, setPreviewError] = useState<string | null>(null);
   const [joinError, setJoinError] = useState<string | null>(null);
   const [joinResult, setJoinResult] = useState<JoinRoomResponse | null>(null);
   const [isJoining, setIsJoining] = useState(false);
@@ -69,10 +50,6 @@ export function App() {
       isOnline: typeof navigator === "undefined" ? true : navigator.onLine,
     }),
   );
-  const [deferredInstallPrompt, setDeferredInstallPrompt] = useState<BeforeInstallPromptEvent | null>(null);
-  const [installMessage, setInstallMessage] = useState<string | null>(null);
-  const [isInstallingApp, setIsInstallingApp] = useState(false);
-  const [isStandaloneApp, setIsStandaloneApp] = useState(() => getStandaloneAppState());
   const hostSecret = useMemo(
     () => (viewState.kind === "home" ? null : loadStoredHostSecret(window.localStorage, viewState.slug)),
     [viewState],
@@ -150,8 +127,27 @@ export function App() {
     setViewState,
     viewState,
   });
-  const previewStreamRef = useRef<MediaStream | null>(null);
-  const previewVideoRef = useRef<HTMLVideoElement | null>(null);
+  const {
+    handleInstallApp,
+    installMessage,
+    isInstallingApp,
+    isStandaloneApp,
+    showInstallPrompt,
+  } = useInstallPrompt();
+  const {
+    clearPreview,
+    handleStartPreview,
+    previewAudioEnabled,
+    previewError,
+    previewState,
+    previewVideoEnabled,
+    previewVideoRef,
+    requestedMedia: previewRequestedMedia,
+    setPreviewAudioEnabled,
+    setPreviewVideoEnabled,
+  } = useDevicePreview({
+    viewState,
+  });
 
   useEffect(() => {
     const handlePopState = () => {
@@ -162,23 +158,6 @@ export function App() {
     return () => {
       window.removeEventListener("popstate", handlePopState);
     };
-  }, []);
-
-  useEffect(() => {
-    setIsStandaloneApp(getStandaloneAppState());
-
-    return attachInstallPromptListeners(window, {
-      onPromptAvailable: (event) => {
-        setDeferredInstallPrompt(event);
-        setInstallMessage("Install LowTime for faster access from your home screen.");
-      },
-      onInstalled: () => {
-        setDeferredInstallPrompt(null);
-        setIsInstallingApp(false);
-        setIsStandaloneApp(true);
-        setInstallMessage("LowTime is installed and ready to launch like an app.");
-      },
-    });
   }, []);
 
   useEffect(() => {
@@ -217,30 +196,7 @@ export function App() {
     setJoinResult(null);
     setDisplayName("");
     setSelectedQualityPreset(DEFAULT_QUALITY_PRESET);
-    setPreviewAudioEnabled(DEFAULT_REQUESTED_MEDIA.audio);
-    setPreviewVideoEnabled(DEFAULT_REQUESTED_MEDIA.video);
-    setPreviewState("idle");
-    setPreviewError(null);
-    stopMediaStream(previewStreamRef.current);
-    previewStreamRef.current = null;
   }, [viewState]);
-
-  useEffect(() => {
-    const videoElement = previewVideoRef.current;
-    const stream = previewStreamRef.current;
-
-    if (videoElement == null) {
-      return;
-    }
-
-    videoElement.srcObject = stream;
-
-    return () => {
-      if (videoElement.srcObject === stream) {
-        videoElement.srcObject = null;
-      }
-    };
-  }, [previewState, previewVideoEnabled]);
 
   async function handleCreateRoom() {
     setIsCreating(true);
@@ -279,26 +235,6 @@ export function App() {
     await navigator.clipboard.writeText(toAbsoluteJoinUrl(createResult.joinUrl));
   }
 
-  async function handleInstallApp() {
-    if (deferredInstallPrompt == null) {
-      return;
-    }
-
-    setIsInstallingApp(true);
-
-    try {
-      const outcome = await promptForInstallation(deferredInstallPrompt);
-      setDeferredInstallPrompt(null);
-      setInstallMessage(
-        outcome === "accepted"
-          ? "Install accepted. Your browser will finish adding LowTime."
-          : "Install dismissed. You can still add LowTime from your browser menu later.",
-      );
-    } finally {
-      setIsInstallingApp(false);
-    }
-  }
-
   function handleOpenRoom() {
     if (createResult == null) {
       return;
@@ -322,7 +258,7 @@ export function App() {
         apiBaseUrl,
         displayName,
         qualityPreset: selectedQualityPreset,
-        requestedMedia: buildRequestedMedia(previewAudioEnabled, previewVideoEnabled),
+        requestedMedia: previewRequestedMedia,
         slug: viewState.slug,
       });
       setJoinResult(payload);
@@ -333,11 +269,10 @@ export function App() {
           displayName: displayName.trim(),
           qualityPreset: selectedQualityPreset,
           transportPreference: payload.transportPreference,
-          requestedMedia: buildRequestedMedia(previewAudioEnabled, previewVideoEnabled),
+          requestedMedia: previewRequestedMedia,
         });
 
-        stopMediaStream(previewStreamRef.current);
-        previewStreamRef.current = null;
+        clearPreview();
         window.history.pushState({}, "", getCallRoute(viewState.slug));
         setViewState(getViewState(window.location.pathname));
       } else if (payload.joinState === "waiting") {
@@ -345,7 +280,7 @@ export function App() {
           requestId: payload.requestId,
           displayName: displayName.trim(),
           qualityPreset: selectedQualityPreset,
-          requestedMedia: buildRequestedMedia(previewAudioEnabled, previewVideoEnabled),
+          requestedMedia: previewRequestedMedia,
         };
 
         saveStoredLobbyRequest(window.sessionStorage, viewState.slug, storedRequest);
@@ -356,30 +291,6 @@ export function App() {
       setJoinError(error instanceof Error ? error.message : "Unable to join room");
     } finally {
       setIsJoining(false);
-    }
-  }
-
-  async function handleStartPreview() {
-    setPreviewState("requesting");
-    setPreviewError(null);
-
-    try {
-      const stream = await startPreviewRequest(buildRequestedMedia(previewAudioEnabled, previewVideoEnabled));
-      stopMediaStream(previewStreamRef.current);
-      previewStreamRef.current = stream;
-      setPreviewState("ready");
-    } catch (error) {
-      stopMediaStream(previewStreamRef.current);
-      previewStreamRef.current = null;
-
-      const name = error instanceof DOMException ? error.name : "";
-      if (name === "NotAllowedError" || name === "SecurityError") {
-        setPreviewState("blocked");
-        setPreviewError("Camera or microphone access was blocked. Adjust browser permissions to preview devices.");
-      } else {
-        setPreviewState("error");
-        setPreviewError(error instanceof Error ? error.message : "Unable to start device preview.");
-      }
     }
   }
 
@@ -491,7 +402,7 @@ export function App() {
       isStandaloneApp={isStandaloneApp}
       installMessage={installMessage}
       shareUrl={createResult ? toAbsoluteJoinUrl(createResult.joinUrl) : null}
-      showInstallPrompt={deferredInstallPrompt != null}
+      showInstallPrompt={showInstallPrompt}
       onCopyLink={handleCopyLink}
       onCreateRoom={handleCreateRoom}
       onInstallApp={handleInstallApp}
@@ -521,11 +432,4 @@ function getNavigatorConnection(): NavigatorConnectionLike | null {
   };
 
   return candidate.connection ?? candidate.mozConnection ?? candidate.webkitConnection ?? null;
-}
-
-function getStandaloneAppState(): boolean {
-  return isPwaInstalled({
-    matchMedia: typeof window.matchMedia === "function" ? window.matchMedia.bind(window) : undefined,
-    navigator,
-  });
 }
